Data_Scraping/Multi_source_scraper/requirements.txt
requests
beautifulsoup4
newspaper3k
youtube-transcript-api
langdetect
yake
lxml